python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
boto3==1.28.64
pydantic==2.4.2
sse-starlette==1.6.5
